Initial commit
Adding two ardino sketches:

*one_axis_balance:
Contains code from the i2cdevlib by jrowberg:
see https://github.com/jrowberg/i2cdevlib

In this sketch the quadcopter is made to balance in one rotational axis using
feedback from the sensor and a PID loop.

A processing sketch is contained to allow the data to be graphed for
analysis.

*single_motor:
This sketch allows the motors to be controlled individually from the serial
port for testing and calibration.