Quadcopter simulator now useable!
Simulation now works in one angular dimension. Lots of work still needs doing
to get it to the stage where all angular dimensions are considered. And then
there's linear motion...

A PID loop has been implemented and the P term has been tested, the other
terms need testing too. This really ought to be done using the graph plotting
stuff written for the arduino, so we need a way of either streaming data
between the programs. A simpler way would be to write the data to a file and
graph it from there...